backend/requirements.txt
flask
flask-smorest
flask-cors
marshmallow
python-dotenv
flask-sqlalchemy
flask-jwt-extended
psycopg2-binary
flask-migrate
